+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 <!-- next-version -->
 
+## v1.14.0 (2025-10-09)
+
+### Fixes
+
+- use time_utils handle time parse and timestamp
+- test
+- use time_utils handle time parse and timestamp
+
+## v1.13.1 (2025-10-06)
+
+### Features
+
+- new timestampe file and new resampler
+- add univ_dic2 and speed up lsr oi fr download and some bug fix
+
+### Fixes
+
+- release workflow file format error
+- release workflow file format error
+
+
+
 ## v1.12.1 (2025-10-02)
 
 ### Fixes